+++ packages/expo-ui/CHANGELOG.md
@@ -10,7 +10,7 @@
 
 ### 💡 Others
 
-- Improved `@expo/ui/swift-ui-primitives` integrations. ([#36937](https://github.com/expo/expo/pull/36937) by [@kudo](https://github.com/kudo))
+- Improved `@expo/ui/swift-ui-primitives` integrations. ([#36937](https://github.com/expo/expo/pull/36937), [#36938](https://github.com/expo/expo/pull/36938) by [@kudo](https://github.com/kudo))
 
 ## 0.1.1-alpha.7 — 2025-04-30